fix: fixed PointMaterial on WebGL1
PointMaterial is not working on the browser that are using previous version of WebGL.
It's because 'GL_OES_standard_derivatives' extension. In WebGL2, this extension has been included in core.
For WebGL1 support, we need to enable it first. We also need to be sure to not enable it on WebGL2,
because it will throw an error.