fix(Text): add missing whitespace option to types
The underlying troika-text-three instance uses the following as
options for the 'whitespace' property:

- 'normal'
- 'nowrap'

Where 'normal' indicates that the 'overflowWrap' property should
be enforced, and 'nowrap' indicates that no wrapping should
occur.

This commit adds the missing 'nowrap' option to the existing prop
types. Note that the extraneous 'overflowWrap' option -- while
unused under the hood -- is remaining such that no breaking
changes are introduced.

fixes issue #1199