fix: `<Select/>` when `!multiple || !box``
When the props `multiple` and `box` are defined, `downed` is being
set through eventListeners set on the window inside an effect.
If they are not set, `downed` before this commit was never set,
which caused `onChange` to never run.

To solve this bug I add a condition `shouldDragSelect = multiple && box`
a conditional event-listeners on `<group/>` in case `!shouldDragSelect`
so that when `shouldDragSelect` `down` is being set through the
event-listeners set inside the effect, and if `!shouldDragSelect` it
is being set whenever one of Select's children is being clicked.