fix: add `shouldDragSelect` to dependencies
Before commit the effect relying on `shouldDragSelect` did not include it in its dependencies.
Neither did the code before this branch, so I am unsure if this was a conscious act or not,
but I believe it could cause a bug where one of the props is changed but event-listeners are not properly cleaned up.

besides this I return `undefined` instead of `null` in case `shouldDragSelect` in `<group/>` event-listeners