feat: add prop to configure canvasRenderingContext2DSettings (#2104)
* refactor: Optimize useSpriteLoader performance

This commit optimizes the performance of the useSpriteLoader function by adding the 'willReadFrequently' flag to the getContext method call. This flag improves the rendering performance when reading from the canvas context. The change ensures that the getContext method is called with the 'willReadFrequently' flag set to true, which can result in faster rendering of sprite textures.

* add space

* refactor: add canvasRenderingContext2DSettings prop

- Added the canvasRenderingContext2DSettings prop to the SpriteAnimator component in the core module.

* document canvasRenderingContext2DSettings prop

* update docs